Standard Schema Project Layout and Tests
This also runs on IBM DFDL - one must comment out an assert that
uses dfdl:checkConstraints (in common.dfdl.xsd), and only one of the
two tests works. The other seems to get a numeric failure on IBM DFDL
perhaps related to DAFFODIL-2139. There are unsignedLong values in this
data that are well beyond the range of a max signedLong.