Improve portability. Fix broken unsignedint64 definition.
The unsignedint64 was defined as a 64-bit long unsignedInt, but that has
a maximum of 32-bits, so changed to extend an unsignedLong.

Changed assertion to avoid IBM DFDL restrictions on assertion placement.
Changed dfdl:checkConstraints to a check for ". eq 0" which is run by
IBM DFDL. 

Changed a test to twoPass because it tolerates but does not regenerate
padding at end of records.